// tshe-thau-loo/src/app/models/activity/general-campaign/general-campaign-edit.model.ts
export interface IGeneralCampaignEdit {
  title: string;
  content: string;
  declaration: string;
  venue: string;
  registrationStartDate: Date;
  registrationStartTime: Date;
  registrationEndDate: Date;
  registrationEndTime: Date;
  startDate: Date;
  startTime: Date;
  endDate: Date;
  endTime: Date;
  limitNumberOfPeople: number;
  enableVerify: boolean;
  enableIdentityConfirmed: boolean;
}